Revert escaping dollar symbol (#193)
Franklin automatically converts `$x$` to `\(x\)` and Pluto to `<span>`
etc, so the best option is to avoid convert `$x$` via KaTeX in
Javascript post processing. Then, dollars do not have to be escaped.
This solves the problem where it is actually not so straightforward in
Javascript to un-escape dollar symbols after running KaTeX.

EDIT: Maybe better would be for PlutoStaticHTML.jl to not escape dollar
symbols, but instead to convert inline math dollar symbols to `\(` and
`\)` like Franklin.jl does.